+++ app/models/grandprix_event.rb
@@ -1,6 +1,0 @@
-class GrandprixEvent < ApplicationRecord
-  has_many :grandprix_entries, dependent: :destroy
-  has_many :skaters, through: :grandprix_entries
-
-  belongs_to :category
-end