+++ config/routes.rb
@@ -27,7 +27,7 @@ Rails.application.routes.draw do
     get :list, on: :collection
   end
 
-  resources :grandprixes, only: :index
+  #resources :grandprixes, only: :index
 
   resources :parsers, only: :index do
     collection do