+++ CHANGELOG.md
@@ -17,6 +17,11 @@
 - spec: update competition details, deviation
 - view: link_to_filter
 - season - minus operator
+- update season (adding to season_from, _to)
+- spec rake: update:grandprix
+- grandprix: dont run if FINAL RESULTS ready
+- spec update: wtt2017 tz 
+- grandprix locale ja
 
 ### new feature